+++ ginga/util/rgb_cms.py
@@ -11,6 +11,7 @@ import hashlib
 import tempfile
 import numpy as np
 
+from ginga import trcalc
 from ginga.misc import Bunch
 from ginga.util.toolbox import PIL_LT_9_1
 
@@ -18,16 +19,20 @@ from . import paths
 
 from PIL import Image
 # How about color management (ICC profile) support?
+have_cms = False
+have_pil_lcms = False
+
 try:
     import PIL.ImageCms as ImageCms
+    have_pil_lcms = True
     have_cms = True
 except ImportError:
-    have_cms = False
+    pass
 
 basedir = paths.ginga_home
 
 # Holds profiles
-profile = {}
+profile_dict = {}
 rendering_intent = 'perceptual'
 intents = dict(perceptual=0)
 
@@ -35,14 +40,14 @@ intents = dict(perceptual=0)
 working_profile = None
 
 # Holds transforms
-icc_transform = {}
+icc_transform_dict = {}
 
 
 class ColorManagerError(Exception):
     pass
 
 
-class ColorManager(object):
+class ColorManager:
 
     def __init__(self, logger):
         self.logger = logger
@@ -50,29 +55,33 @@ class ColorManager(object):
     def can_profile(self):
         return have_cms
 
-    def profile_to_working_pil(self, image_pil, kwds, intent=None):
-
-        image = image_pil
+    def profile_to_working_numpy(self, image_np, kwds, intent=None):
 
         # If we have a working color profile then handle any embedded
         # profile or color space information, if possible
         if not have_cms:
             self.logger.info(
                 "No CMS is installed; leaving image unprofiled.")
-            return image
+            return image_np
 
         if not have_profile(working_profile):
             self.logger.info(
                 "No working profile defined; leaving image unprofiled.")
-            return image
+            return image_np
 
-        out_profile = profile[working_profile].name
+        out_profile = profile_dict[working_profile].name
 
-        if not os.path.exists(profile[out_profile].path):
+        if not os.path.exists(profile_dict[out_profile].path):
             self.logger.info(
                 "Working profile '%s' (%s) not found; leaving image "
-                "unprofiled." % (out_profile, profile[out_profile].path))
-            return image
+                "unprofiled." % (out_profile, profile_dict[out_profile].path))
+            return image_np
+
+        if image_np.dtype != np.uint8:
+            ## image_np = trcalc.array_convert(image_np, np.dtype(np.uint8))
+            self.logger.info(
+                "Cannot profile >8 bpp images; leaving image unprofiled.")
+            return image_np
 
         if intent is None:
             intent = rendering_intent
@@ -80,16 +89,21 @@ class ColorManager(object):
         # Assume sRGB image, unless we learn to the contrary
         in_profile = 'sRGB'
         try:
-            if 'icc_profile' in image.info:
-                self.logger.debug("image has embedded color profile")
-                buf_profile = image.info['icc_profile']
+            if 'icc_profile' in kwds:
+                self.logger.info("image has embedded color profile")
+                buf_profile = kwds['icc_profile']
+                # NOTE: is there a way to get a name for this embedded profile?
+                # make up a unique name
+                in_profile = hashlib.md5(buf_profile).hexdigest()  # nosec
 
                 # Write out embedded profile (if needed)
-                prof_md5 = hashlib.md5(buf_profile).hexdigest()  # nosec
-                _fd, in_profile = tempfile.mkstemp(suffix=".icc",
-                                                   prefix="_image_{}_".format(prof_md5))
-                with os.fdopen(_fd, 'wb') as icc_f:
-                    icc_f.write(buf_profile)
+                if in_profile not in profile_dict:
+                    _fd, path = tempfile.mkstemp(suffix=".icc",
+                                                 prefix="_image_{}_".format(in_profile))
+                    with os.fdopen(_fd, 'wb') as icc_f:
+                        icc_f.write(buf_profile)
+                    profile_dict[in_profile] = Bunch.Bunch(name=in_profile,
+                                                           path=os.path.abspath(path))
 
             # see if there is any EXIF tag about the colorspace
             elif 'ColorSpace' in kwds:
@@ -110,61 +124,59 @@ class ColorManager(object):
                 else:
                     self.logger.debug("no color space metadata, assuming this is an sRGB image")
 
-            # if we have a valid profile, try the conversion
-            tr_key = get_transform_key(in_profile, out_profile, intent,
-                                       None, None, 0)
-
             # check if image has an alpha channel; if so we need to remove
             # it before ICC transform and tack it back on afterwards
-            alpha = None
-            if 'a' in image.mode.lower():
-                channels = image.split()
-                alpha = channels[-1]
-                image = Image.merge("RGB", channels[:-1])
+            image_rgb, alpha = image_np, None
+            if 'A' in trcalc.guess_order(image_np.shape):
+                image_rgb, alpha = trcalc.remove_alpha(image_np)
 
-            if tr_key in icc_transform:
-                # We have an in-core transform already for this (faster)
-                image = convert_profile_pil_transform(image,
-                                                      icc_transform[tr_key],
-                                                      inPlace=True)
-            else:
-                # Convert using profiles on disk (slower)
-                if in_profile in profile:
-                    in_profile = profile[in_profile].path
-                image = convert_profile_pil(image, in_profile,
-                                            profile[out_profile].path,
-                                            intent)
+            if have_pil_lcms:
+                # fallback to LCMS bundled with pillow, if available
+                # if we have a valid profile, try the conversion
+                tr_key = get_transform_key(in_profile, out_profile, intent,
+                                           None, None, 0)
+
+                # convert numpy image to PIL format
+                image_pil = to_image(image_rgb)
+
+                if tr_key in icc_transform_dict:
+                    # We have an in-core transform already for this (faster)
+                    image_pil = convert_profile_pil_transform(image_pil,
+                                                              icc_transform_dict[tr_key],
+                                                              inPlace=True)
+                else:
+                    # Convert using profiles on disk (slower)
+                    if in_profile in profile_dict:
+                        in_profile = profile_dict[in_profile].path
+                        image_pil = convert_profile_pil(image_pil, in_profile,
+                                                        profile_dict[out_profile].path,
+                                                        intent)
+
+                # convert PIL image to numpy format
+                image_rgb = from_image(image_pil)
 
             # reattach alpha channel if there was one
             if alpha is not None:
-                channels = list(image.split())
-                channels.append(alpha)
-                image = Image.merge("RGBA", channels)
+                image_rgb = trcalc.add_alpha(image_rgb, alpha=alpha)
+            image_np = image_rgb
 
             self.logger.info("converted from profile (%s) to profile (%s)" % (
-                in_profile, profile[out_profile].name))
+                in_profile, profile_dict[out_profile].name))
 
         except Exception as e:
             self.logger.error("Error converting from embedded color profile: {!r}".format(e),
                               exc_info=True)
             self.logger.warning("Leaving image unprofiled.")
 
-        return image
+        return image_np
 
-    def profile_to_working_numpy(self, image_np, kwds, intent=None):
-
-        if image_np.dtype != np.uint8:
-            self.logger.info(
-                "Cannot profile >8 bpp images; leaving image unprofiled.")
-            return image_np
-
-        image_in = to_image(image_np)
-        image_out = self.profile_to_working_pil(image_in, kwds,
-                                                intent=intent)
-        return from_image(image_out)
-
+    def profile_to_working_pil(self, image_pil, kwds, intent=None):
+        image_np = from_image(image_pil)
+        image_np = self.profile_to_working_numpy(image_np, kwds, intent=intent)
+        return to_image(image_np)
 
 # --- Color Management conversion functions ---
+
 
 def to_image(image_np, flip_y=True):
     if flip_y:
@@ -233,7 +245,7 @@ def get_transform_key(from_name, to_name, to_intent, proof_name,
 def get_transform(from_name, to_name, to_intent='perceptual',
                   proof_name=None, proof_intent=None,
                   use_black_pt=False):
-    global icc_transform
+    global icc_transform_dict
 
     if not have_cms:
         return ColorManagerError("Either pillow is not installed, or there is "
@@ -256,29 +268,30 @@ def get_transform(from_name, to_name, to_intent='perceptual',
                             proof_intent, flags)
 
     try:
-        output_transform = icc_transform[key]
+        output_transform = icc_transform_dict[key]
 
     except KeyError:
         # try to build transform on the fly
         try:
             if proof_name is not None:
                 output_transform = ImageCms.buildProofTransform(
-                    profile[from_name].path,
-                    profile[to_name].path,
-                    profile[proof_name].path,
+                    profile_dict[from_name].path,
+                    profile_dict[to_name].path,
+                    profile_dict[proof_name].path,
                     'RGB', 'RGB',
                     renderingIntent=intents[to_intent],
                     proofRenderingIntent=intents[proof_intent],
                     flags=flags)
             else:
                 output_transform = ImageCms.buildTransform(
-                    profile[from_name].path,
-                    profile[to_name].path,
+                    profile_dict[from_name].path,
+                    profile_dict[to_name].path,
                     'RGB', 'RGB',
                     renderingIntent=intents[to_intent],
                     flags=flags)
 
-            icc_transform[key] = output_transform
+            # cache it so we don't have to build it later
+            icc_transform_dict[key] = output_transform
 
         except Exception as e:
             raise ColorManagerError("Failed to build profile transform: {!r}".format(e))
@@ -291,6 +304,19 @@ def convert_profile_fromto(image_np, from_name, to_name,
                            proof_name=None, proof_intent=None,
                            use_black_pt=False, logger=None):
 
+    if image_np.dtype != np.uint8:
+        ## image_np = trcalc.array_convert(image_np, np.dtype(np.uint8))
+        if logger is not None:
+            logger.info(
+                "Cannot profile >8 bpp images; leaving image unprofiled.")
+        return image_np
+
+    alpha = None
+    ht, wd, dp = image_np.shape
+    if dp > 3:
+        # color profile conversion does not handle an alpha layer
+        image_np, alpha = trcalc.remove_alpha(image_np)
+
     try:
         output_transform = get_transform(from_name, to_name,
                                          to_intent=to_intent,
@@ -299,6 +325,11 @@ def convert_profile_fromto(image_np, from_name, to_name,
                                          use_black_pt=use_black_pt)
 
         out_np = convert_profile_numpy_transform(image_np, output_transform)
+
+        if alpha is not None:
+            # reattach alpha layer if there was one
+            out_np = trcalc.add_alpha(out_np, alpha)
+
         return out_np
 
     except Exception as e:
@@ -328,10 +359,10 @@ glob_pat = os.path.join(basedir, "profiles", "*.icc")
 for path in glob.glob(glob_pat):
     dirname, filename = os.path.split(path)
     profname, ext = os.path.splitext(filename)
-    profile[profname] = Bunch.Bunch(name=profname,
-                                    path=os.path.abspath(path))
+    profile_dict[profname] = Bunch.Bunch(name=profname,
+                                         path=os.path.abspath(path))
 
-if have_cms:
+if have_pil_lcms:
     if PIL_LT_9_1:
         d = dict(absolute_colorimetric=ImageCms.INTENT_ABSOLUTE_COLORIMETRIC,
                  perceptual=ImageCms.INTENT_PERCEPTUAL,
@@ -348,11 +379,11 @@ if have_cms:
 
 
 def have_profile(name):
-    return name in profile.keys()
+    return name in profile_dict.keys()
 
 
 def get_profiles():
-    names = list(profile.keys())
+    names = list(profile_dict.keys())
     names.sort()
     return names
 
@@ -364,7 +395,5 @@ def get_intents():
 
 
 def set_profile_alias(alias, profname):
-    global profile
-    profile[alias] = profile[profname]
-
-# END
+    global profile_dict
+    profile_dict[alias] = profile_dict[profname]