+++ doc/WhatsNew.rst
@@ -13,6 +13,7 @@ Ver 3.4.0 (unreleased)
 - fixed a bug in ingesting the metadata (header) of RGB images
 - added support for backends pyqt6 and pyside6; removed support for
   pyqt4 and pyside
+- fix for a bug in ICC profiling with temp file creation
 
 Ver 3.3.0 (2022-02-16)
 ======================